Bring back self argument for ApogeePred class
Also add some comments and type hints

+++ airbrakes/data_handling/processed_data_packet.py
@@ -1,6 +1,7 @@
 """Module for describing the data packet for the processed IMU data"""
 
 import msgspec
+import numpy as np
 
 
 class ProcessedDataPacket(msgspec.Struct):
@@ -9,7 +10,7 @@ class ProcessedDataPacket(msgspec.Struct):
     data.
     """
 
-    current_altitude: float
-    vertical_velocity: float
-    vertical_acceleration: float
-    time_since_last_data_point: float  # dt is the time difference between the current and previous data point
+    current_altitude: np.float64  # This is the zeroed-out altitude of the rocket.
+    vertical_velocity: np.float64  # This is the velocity of the rocket, in the upward axis (whichever way is up)
+    vertical_acceleration: np.float64  # This is the rotated compensated acceleration of the vertical axis
+    time_since_last_data_point: np.float64  # dt is the time difference between the current and previous data point
